# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FitTrack API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/sync": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Sync Changes",
        "operationId": "sync_changes_api_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/google-fit/sync": {
      "post": {
        "tags": [
          "google-fit"
        ],
        "summary": "Sync Google Fit",
        "operationId": "sync_google_fit_api_google_fit_sync_post",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Date to sync in YYYY-MM-DD",
              "title": "Date"
            },
            "description": "Date to sync in YYYY-MM-DD"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/google-calendar/sync": {
      "post": {
        "tags": [
          "google-calendar"
        ],
        "summary": "Sync Google Calendar",
        "operationId": "sync_google_calendar_api_google_calendar_sync_post",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Date to sync in YYYY-MM-DD",
              "title": "Date"
            },
            "description": "Date to sync in YYYY-MM-DD"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SyncChangeItem": {
        "properties": {
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "entity_id": {
            "type": "string",
            "format": "uuid",
            "title": "Entity Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "client_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Client Timestamp"
          }
        },
        "type": "object",
        "required": [
          "entity_type",
          "entity_id",
          "action",
          "payload",
          "client_timestamp"
        ],
        "title": "SyncChangeItem"
      },
      "SyncConflictItem": {
        "properties": {
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "entity_id": {
            "type": "string",
            "format": "uuid",
            "title": "Entity Id"
          },
          "client_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Client Payload"
          },
          "server_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Server Payload"
          },
          "client_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Client Timestamp"
          },
          "server_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Server Timestamp"
          }
        },
        "type": "object",
        "required": [
          "entity_type",
          "entity_id",
          "client_payload",
          "server_payload",
          "client_timestamp",
          "server_timestamp"
        ],
        "title": "SyncConflictItem"
      },
      "SyncOperationResult": {
        "properties": {
          "change_index": {
            "type": "integer",
            "title": "Change Index"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "entity_id": {
            "type": "string",
            "format": "uuid",
            "title": "Entity Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "applied",
              "duplicate",
              "conflict",
              "validation_error"
            ],
            "title": "Status"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "change_index",
          "entity_type",
          "entity_id",
          "status"
        ],
        "title": "SyncOperationResult"
      },
      "SyncRequest": {
        "properties": {
          "last_sync": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync"
          },
          "changes": {
            "items": {
              "$ref": "#/components/schemas/SyncChangeItem"
            },
            "type": "array",
            "title": "Changes",
            "default": []
          },
          "client_id": {
            "type": "string",
            "title": "Client Id",
            "default": "default"
          }
        },
        "type": "object",
        "title": "SyncRequest"
      },
      "SyncResponse": {
        "properties": {
          "server_changes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Server Changes",
            "default": []
          },
          "conflicts": {
            "items": {
              "$ref": "#/components/schemas/SyncConflictItem"
            },
            "type": "array",
            "title": "Conflicts",
            "default": []
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SyncOperationResult"
            },
            "type": "array",
            "title": "Results",
            "default": []
          },
          "sync_token": {
            "type": "string",
            "format": "date-time",
            "title": "Sync Token"
          }
        },
        "type": "object",
        "required": [
          "sync_token"
        ],
        "title": "SyncResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}